Debt tracker (#23)
* Add .env options for debt tracking

* Add loaded debt parameters to Config

* Create strategies skeleton

* Create standard strategy

* Add splitwise balance strategy

* Remove redundant option

* Import and use strategy classes

* Use paid amount instead of owed on flag

* Modify processExpense to use correct strategy

* Actually use paid amount on flag

* Add strategy testing skeleton

* Remove redundant option from main

* Convert str to float for comparison

* minor strategy tests changes

* test bugfix

* Fix processExpense logic

* Remove redundant patch

* minor bugfixes

* Remove use_payed_amount option from processExpense

Handle it as one of the arguments to applyExpenseAmountToTransaction

* Use applyExpenseAmountToTransaction in SW balance strategy

* Correctly make a deposit to the balance account

Make a deposit to the SW balance account of amount "balance" when an expense entails people owe me money, instead of a transfer from the paying account to the SW balance, which doesn't make sense.

* bug: handle deposit type txns

If type is "deposit", the asset account is the *destination*, not the *source*.

* convert str to float for taking negative

* handle expenses by others

If I payed nothing on an expense, don't make a transaction from my real accounts, only from the balance account.

* unify comment style

* refactor for clarity

explicitly handle expenses where I paid nothing but I still owe something, and the edge case where an expense was recorded but I paid my whole share.

* flatten dictionary to comply with signature

* clarify amount application method usage

 Remove ability to pass ExpenseUser to get amount, pass amount explicitly. Rename method to clarify usage.

* minor refactor

* fix test_sw_balance_strategy

Correctly use mock_apply_transaction_amount

* Make test robust to trailing zeros

* Disable splitwise balance feature in test

* Extend README to describe debt tracking